Display current board correctly in UI.
This wasn't working before because I was calling a new instance
of the game board from the displayController, and not using the
game board being used in the gameController.  By returning the game
board in the gameController function, I can reference it in the
displayController, thus making sure all components are working on the
same board.